Deep link and start deep link improvements.
Allow access to deeplinks before window load completes.

With this change, deep links are available immediately, as soon as
they are received.  The delay until after window load completes is
removed.

This also brings back the initialDeepLink property for polling access
to deep link values.  The initialDeepLink value will return updated
deep link values as long as onDeepLink listeners are not added.

If the initialDeepLink value was not accessed, the first added
onDeepLink listener will receive it synchronously.

This ensures that deep links are reported in the same way as initial
deep links.

The initialDeepLink property value will be cleared when a new
document is loaded, unless its value has not been accessed.

As a result, when page redirects are used, the first page that
actually accesses the deep link receives it, and none of the pages
loaded afterwards receive it.

Extensive black box tests are added for the various combinations of
receiving deep links, and navigating/redirecting.

This also removes the previously deprecated ApplicationEvent.

This fixes b/162335268

b/140013592
b/145948517
b/157996837

This fixes b/147225366
This should improve b/145923478

Change-Id: I74a9c140b67a78ec14ee3268174396e0eecd497b

Former-commit-id: 621098474b7806a8d784e32c7f2f277877261c5e