Provide default version when reading manifest fails.
This change adds |kDefaultManifestVersion|, with the value |"1.0.0"|, to
be used as the default version returned when reading the manifest fails
and the current slot is the system image.  This is useful for when the
manifest is malformed, or missing, in the system imaage as it allows an
update to still be downloaded.

Change-Id: Iab3c723ce50d1af88b43462f1d2694d6045b1654
(cherry picked from commit 7087212c0492e365da2a1ca89a9128d393bcba63)

Former-commit-id: 01a4b9d4e995912548bcc68e923aa5969418f1d3