Provide default version when reading manifest fails.
This change adds |kDefaultManifestVersion|, with the value |"1.0.0"|, to
be used as the default version returned when reading the manifest fails
and the current slot is the system image.  This is useful for when the
manifest is malformed, or missing, in the system imaage as it allows an
update to still be downloaded.

Change-Id: Iab3c723ce50d1af88b43462f1d2694d6045b1654

Former-commit-id: eedcfd2bfef37fd06911f02fe117d2322314f062