Accept java crash when calling beforeSuspend at stop
Calling beforeSusupend when Android stop signal was received can result
in a JNI crash. This beforeSuspend call was introduced to kill
MediaSession earlier to fix avoid showing users media content 1 second
after suspend. But this beforeSuspend call is causing considerable
amount of crashes in production and I think we need to think of a way to
either avoid the crash or ignore the crash. This CL chose to implement
the later.

b/165052794
b/131170502

Change-Id: Ic5ed5a84addab3b21aeadac8efadf8ab42c0b828

Former-commit-id: 7b36d9c166f8cce0fd2ad80236d97e65628717b1