Add docker container for building with clang 3.6
Containerize the build for linux-x64x11-clang-3-6 configuration by using
a different BASE_OS that is capable of installing clang 3.6 from the
repos. This is necessary because of the removal of libstdc++-7-dev from
the rodete repos which makes the normal compilation of the clang 3.6
toolchain impossible in-situ.

The new container is based off ubuntu xenial and is built with:
```
docker-compose up linux-x64x11-clang-3-6
```

b/167074536

Change-Id: I05c4221f1b5a77c0d06f38928608c096f69be1ee

Former-commit-id: 7fe6c2d7e992e6acb7007752fda87bb0289fa865