Propagate focus to the nearest ancestor with a hybrid nav focus item
This change allows the web app to focus one DOM html element while
making an ancestor the focus in the hybrid nav system. This simplifies
situations where the desired DOM focus element is not large enough
to encapsulate the focus area, so a parent DOM element needs to be made
the focus item.

Also ensure that only enabled hybrid nav items can be focused.

b/166516925

Change-Id: I3888c21b5b23613b17f75b5091c335e3f914cbb7

Former-commit-id: 43d48d5d4bfb366ab677e409feeadfbb2f154365