Skip "out of storage" test on BuildBot.
The "out of storage" test requires sudo to ensure we have a temporary
filesystem mounted at ~/.cobalt_storage.tmpfs which requires sudo. We
will not have this on buildbot and cannot be certain that the filesystem
will exist.

This change also includes fixes and improvements to how the temporary
filesystem is mounted.

b/158488347
b/158485220

Change-Id: I66b9605344eb5b6407c5c5d9894c51b71fad235c

Former-commit-id: 9ce1b64b21459303a0a04a7f674a6362280eebcc