[Android] Refine MediaCodec width and height query
The current implementation separately queries the maximum frame height
and width independently of each other. This will cause a playback issue
if the maximum possible frame height cannot support the maximum frame
width, and vice versa. This change adds a check to see if the maxHeight
and maxWidth can be supported together, and queries multiple standard
resolutions as a fall back.

This also edits an API check to limit 8k only to devices using API level
29 and above.

b/168741143

Change-Id: I03a2cfc5e788579b3e6bf3dd24115b0dd5352343

Former-commit-id: 66217bb36acd85b49936e86c4883285273f1c9f8