Minor Evergreen automated test improvements.
This change updates the log used to signal when an update has been
successfully installed from a channel. Previously, this would be logged
both when we were pending a restart and when we were up to date, but
this is now only logged when we are pending an update. This allows us to
error out earlier.

b/170301033
b/158485220

Change-Id: Ia80294ccc16db3a5310787a55c7bf7dec9e7efe8

Former-commit-id: c8f619516eab613dc3627a1d708287bf98b53370