Write 10s of audio into SbPlayer for preroll
We used to only write up to 1s audio into SbPlayer in advance. But for
wireless devices, it may require more audio data to preroll. We now
buffer up to 10s audio data at the start of the playback or after seek
to make sure the audio output device won't be starving.

b/169337174

Change-Id: I0d24fc33ae2faf233676ae530ed6b22480f4a7c4

Former-commit-id: 4d1a50c7bee8155ed6323ad16f3381c5168838d6