Clean up the configuration_public.h files.
This change moves various SB_HAS_* macros around to be in the
appropriate sections in the configuration_public.h files. This change
introduces two sections, "Platform Specific Configuration" and "Platform
Specific Quirks", to accompany "Platform Specific Audits". The intention
is to have a place for things that actually audit the platform, things
that are not quirks but are intuitive configurations for a platform, and
for unintuitive features of a platform.

b/74087760

Change-Id: I8e19534e494ed06597841313ab26fea6d1f5fece

Former-commit-id: 32a0553cd0a26a162b1d7cef334135804d19b721