Fix windows dockerfiles for better image layering
Dockerfiles need to be refactored into an ordering that places all the
long build-steps in the beginning. This allows us to make changes in
different layers of the final image without triggering a long rebuild.

In this CL, we fix the layering so that the visual studio installation
is factored out into its own base image. Normally this build can take
approximate 30-40 minutes. This image can now be cached, and never needs
to be rebuilt unless we explicitly need to change the visual studio
installation.

b/171313071

Change-Id: I3b1b077a03059dbd3427d6267a7ac71b75d34a6f

Former-commit-id: d92ecc34c1f0f5e72a9648e8d349e69d7a20d66b