Restrict xb1 builds to fewer processes for linking
XB1 builds will use the maximum number of concurrent jobs for linking
however this will cause the builders to run out of memory when run with
the /Z7 compilation flag (that stops the creation of per-object PDB
files).

Ninja allows you to restrict the max number of concurrent processes used
for linking, so this future-proofs it for when we update the PDB
compilation flags.

https://ninja-build.org/manual.html#ref_pool

b/170408484

Change-Id: Ibfc0bbc8743182b26a2974677e43f63e25954e70

Former-commit-id: 91312052f5600628d474c5b71507d3d2d2bd715e