Limit exe and shared lib linking processes for xb1
In a previous CL, static library linking was limited, but executables
and shared libraries was not set. As such we reach OOM errors when
linking for these targets, when compiled with /Z7 msvc flag.

This CL limits the pool of available linking processes for these targets
as well for xb1 builds.

https://ninja-build.org/manual.html#ref_pool

b/170408484

Change-Id: Ie72ef3b3ab814f402d1f1a31225c462ada27ebfe

Former-commit-id: a1e396300147f316074d44b8152313a26cc91024