Refine HTMLMediaElement::SetMediaKeys()
Reset |media_keys_| before CreateMediaPlayer() is called. |media_keys_|
is used in CreateMediaPlayer(). Without resetting it to null, it will
create a player with the old |media_keys_|, which may make the new
player still running with the old drm system even if |media_keys| is
null, or may do redundant work to reset the drm system of the newly
created player.

b/174090554

Change-Id: I6ed81cb41f16794baddee40b7c3142dca0c7c03a

Former-commit-id: 87c464557e34db02ff82b3b26aa7b945a122f398