Deprecate windowManager.getDefaultDisplay()
The android application context does not have a Display attached to
it, so when it is called on getDisplay(), it will throw an exception
causing the app to crash.

NativeActivity is a subclass of Context, it has the Display attached
to it. We need the Display to get the refresh rate, supported Hdr types
and its size.

When a NativeActivity starts, we get the display from the activity and
cached it in DisplayUtil.

Display Size is moved out of the constructor of ArtworkLoader, we
won't know the display size at that time.

b/168803818

Change-Id: I0e16b82ab31eb60a44ca7dd8c891928d01c64964
(cherry picked from commit a45295f271c83531748ffb65e8b4275ce7228c6a)

Former-commit-id: a0b87cf5f7177e954c969be515d9aebe5f9804c0