[linux] Enable vp9 hdr playback
Instead of depending on the system `libvpx` library that may not have
highbitdepth enabled, Cobalt linux-x64x11 now uses a `libvpx.a` built
from `libvpx` trunk.  The shell script to build the `libvpx.a` is also
included so we can update `libvpx.a` in future as needed.

The CL also includes the following changes:
1. Add `libraries.gyp` to `starboard/linux/x64x11/shared/` and use it
for all linux-x64x11 platforms, and put the new `libvpx.a` into it.
Adding the new `libvpx.a` into the `gyp_configuration.gypi` in
`starboard/linux/x64x11/shared` won't work as it is not used by the
`linux-x64x11-clang` and `linux-x64x11-gcc` builds.
2. Modified `clangxx.py` in the abstract toolchain to move the
dependency libraries to the end of the library lists, to conform with
the order requirements of the linker.
3. Removed `third-party/libvpx/libvpx.gyp` as we no longer intend to
build libvpx on any platforms within Cobalt.  Currently platforms on
different libvpx versions may share the same libvpx header files, which
should be ok as the interface is usually stable.  We can either bring up
`libvpx.a` on all platforms to the same version in future, or move
include files into the platform folders when necessary.
4. Removed `third-party/vpx_includes` as they are no longer needed.

Change-Id: I4669c6fb61b909e2d8e3e715511c7302cafe5192

Former-commit-id: 3e8dd9d2384178ea6682e707740c39c1b265fcab