Fix for message pump null pointer read.
Don't try to do work if there is no delegate anymore.  There was a
DCHECK(), but those are stripped in gold builds, resulting in
avoidable crashes on exit when the null pointer is dereferenced.

On UWP, ensure the timer_event_map_ is cleared before the object is
destroyed, to avoid a race when a timer is still active when the
application exits.

This fixes b/175222607

Change-Id: If963a65cba87b0f6eb1b91da72c6ea692567e372

Former-commit-id: d79984a0fd7b962269ce78fbff707836e838af76