[XB1] Add support for compacted texture layout : starboard
Update for AV1 decode target.
This is an adaptation of the video decoder's memory optimization
that consists in decreased texture size. That decrease is possible
due to more efficient pixel data layout of a texture. Instead of
storing each pixel using 16 bits where only 10 of them are used
and 6 are an empty gap, the optimized texture stores triples of
10bit pixel with only 2 bits of unused gaps.

b/148439807

Change-Id: I46cd95c559f0348484a4638cf85aec6fffb9c888

Former-commit-id: af7dc63ffda1da2232fcb3df0dcb195a5f6a1ad2