[android] Refine tunnel mode pipeline
1. Set `useAutoSyncHeaderWrite` in AudioTrackBridge.java to true, to
enable auto sync header writing by default.  Note that this was set to
false to test the manual sync header writing before the last commit, and
failed to revert before committing.
2. Check `media_decoder_` is not null before use in VideoDecoder.  This
variable can be set to null in VideoDecoder::OnSurfaceDestroyed() when
the surface is destroyed, typically happens when the app is suspended.

b/134961794

Change-Id: I9f291df9e1f06e8d391a628140e0fb93c6bc85b0

Former-commit-id: 084ca8abb2c931fdff6f571c5133f5f6f3517038