[android] Force secure playback to enable tunnel mode
On some platforms tunnel mode is only supported in the secure pipeline.
To allow clear videos being played under tunnel mode, we create a dummy
drm system in the VideoDecoder to force creating the secure pipeline.

We may expose this to the web app so it can be configured at run time.
If it turns out to be useful to allow enforcing of the secure pipeline
at run time, this can be refactored in a platform independent way.

b/171146674

Change-Id: Iac166d8745dd9a6d52de99f49ac63efb628091c1

Former-commit-id: 0c061a9c4299320be025d55e3a632e476a73943d