Conditionally enable PDBs for docker or CI builds
Previously, it was unilaterally decided to disbale the generation of PDB
files in order to allow for containerized builds in docker.

However this caused a massive 3x slowdown in the build times for the xb1
nightly builds (which showcases bloated linker times).

As a compromise, in this CL we check for whether the build is dockerized
and based on that we can enable or disable PDB generation. Otherwise by
default it is turned on. This allows the ganeti worker builds to
function as normal.

b/176852468
b/169692017

Change-Id: Ie56147eae4297edd14639cae3bc9818e584c5189

Former-commit-id: eb73f27a182671ff8f3163521674cf66aaa3ee7b