Replace glTexParameterf with glTexParameteri.
This change updates calls to glTexParameterf to instead use
glTexParameteri for:
 - cobalt/renderer/backeng/egl/framebuffer.cc
 - starboard/shared/blittergles/blitter_surface.cc

While other implementations have both functions implemented, Glimp does
not have glTexParameterf implemented. Calling this function could lead
to indexing a mipmap that does not exist, and thus crashing the GPU.
This may be related to b/170325376, but since we have not been able to
reproduce this issue it is still not clear.

b/170325376

Change-Id: Ib299b407de7c4879a49a18c8b5f15bdb3daaa60e

Former-commit-id: e8fb9e4fe957c38866113802940398d9f2017986