Fix hybrid nav misalignment
In some cases, the hybrid nav item does not align with the
corresponding DOM element. This looks to be due to a bug in
Box::GetMarginBoxTransformFromContainingBlock() not accounting
for the content box vs padding box for the containing block's
descendant.

Work around the issue with this bug in the hybrid nav code
for now. This avoids possible knock-on effects in layout and
other code.

b/178261998

Change-Id: Ibcba3926c2669ab014b730b0540a26adf770b4c6
(cherry picked from commit ec691cd3bf1427c90c36d8ea3b676b3c684782bf)

Former-commit-id: a456a1d5935017cd220b94b05cbbbe9a141e2d24