Fix SB_LOG_IF(FATAL, ...) operator precedence error.
The && operator is evaluated before the ternary operator and this caused
an issue with using SB_LOG_IF(FATAL, ...) in gold builds. This issue
caused problems because the && in SB_LOG_IF(...) is evaluated using the
|false| at the end of SB_LOG_IS_ON(...) and the value of the |condition|
provided to SB_LOG_IF(...) *before* the ternary in SB_LOG_IS_ON(...) is
evaluated.

b/178829047

Change-Id: I2674e07a8df9d3ce294579bdd568671b54083181

Former-commit-id: a87f74c003f762032aa79b42d45195b3dd459f55