[android] Enable toggling AudioRouting API by mime
The Android AudioRouting API is currently used to detect when audio is
routed to a new device. As this API was recently introduced to Cobalt,
this change allows this API to be toggled by specifying
"enableaudiorouting=true/false" in the audio mime. If not specified,
AudioRouting is enabled by default.

This change also renames AudioRendererSinkTunneled to
AudioRendererSinkAndroid, and refactors it to be used for both tunneled
and non-tunneled playback.

b/176433002
b/178411853

Change-Id: Id619c46a4141dc1047e000894ed55be915216379
(cherry picked from commit a77ab474c5ff8f0ed2e61172e7378c0fedcb180a)

Former-commit-id: f4fec3be26779e94c18c42ea71d2325a53ae02d9