Update Docker docsite tooling to generate deploy directory.
This change updates the Docker docsite tooling to be able to generate an
easy-to-use deploy directory, resulting in all of the files needed to
update cobalt.dev being generated appropriately from one command.

This change updates the default service on AppEngine to be deployed to
to be the staging service.

b/158002843

Change-Id: I544eebf7ba024d9d24eab3f367ad06912592f4fc
(cherry picked from commit 2667129f609310f96ee58282e306a21bb32410b6)

Former-commit-id: 9f6ff1db032238b3e2e977e64e1b5a9369d58120