Update Docker docsite tooling to generate deploy directory.
This change updates the Docker docsite tooling to be able to generate an
easy-to-use deploy directory, resulting in all of the files needed to
update cobalt.dev being generated appropriately from one command.

This change updates the default service on AppEngine to be deployed to
to be the staging service.

b/158002843

Change-Id: I544eebf7ba024d9d24eab3f367ad06912592f4fc

Former-commit-id: b51a763e66eab0124ed11eb018cbe506ef3d1c68