Add back support of artificial delay on network requests.
This reverts commit 10b8667ff03ca39d03ec947a18063a781e3de63d,
which reverts commit 943602e05bcf81436de4618e06a94a6822ef0034
to implement the support of artificial delay on network requests.

Reason for revert: Fixing the initial mistake that required the
revert.

Moved the definitions of kMaxNetworkDelay and kMaxNetworkDelayHelp in
switches.cc into the ENABLE_DEBUG_COMMAND_LINE_SWITCHES block.

Original Commit Message for the feature:
Add support of artificial delay on network requests

Network requests like XHR are often asynchronous.
Some webapps may accidentally rely on the order of responses of two
network requests to work correctly, and may break when the network
responses are received in a different order. This happens with the
conformance tests as shown in bug b/171854787. This may also happen
to any webapp in production environment.

The feature in this CL adds a new cmd line parameter called
"max_network_delay", which adds a small random delay to the beginning
of XHRs. A random delay time between 0 and the time specified will be
added prior to each XHR's start.

The parameter is specified in microseconds.

b/173450618

Change-Id: Id0ebd0a12712f903f9fd82f3fa1f89ed2f140d46

Former-commit-id: 9099da66b106d9bd47ef6632c5da88208ddc5236