[ICU] Regenerate data files for ICU 68
For ICU 56, Cobalt trimmed its ICU data in the same manner
that Chromium trimmed its ICU data for its Android build
(not for the platform specifications, but simply because that
represented a leaner subset of features/locales, which is more
suitable for Cobalt). This data generated in this CL for ICU 68
continues to trim based on this subset of features; with the
introduction of the ICU data build tool in ICU 64, we can use
the android_small.json filter file provided by Chromium to
make these modifications (instead of having to run scripts to
directly modify resource and build files).

In addition, due to the vast increase in number in regional variants
introduced by the CLDR locale data that ICU uses, we also need to
explicitly specify only the regional variants that we wish to support.
The cobalt.json filter file is based on android_small.json, with these
additional modifications.

We continue to use individual files for memory usage benefits since
not all platforms support mmap.

Please see b/182403146 for further details and followups

Change-Id: Iab7c780515d78842b5b483f39c6b7d578daa6fa9

Former-commit-id: 38978bef85a5ae7e815ebe5ec37341b2de20397d