Dispatch ended event from AudioBufferSourceNode
AudioBufferSourceNode now dispatches an ended event when the audio
finishes playing.
The included web-audio-demo/web-audio-demo.html script plays a few
audio buffer source nodes, printing to the console when their ended
event is handled.

Note: The audio is only audible at the beginning of the 1st
AudioBufferSourceNode playback. There seems to be another bug with
either the WebAudioAPI implementation, or this script's use of it.
This is tangential to the ended event, and should be investigated
and/or fixed.

b/173248525

Change-Id: I3d6a3134abd7d1d548ad3bead9425d6190fa103c

Former-commit-id: 915c9fe2fdef21cfd5cdf729779bcea9f9c027df