[android] Remove byte order on hdrStaticInfo
The byte order was set in the Chromium implementation but not in
ExoPlayer.  It seems that once the byte order was enforced, there are
display issues on some devices (like Dolby Vision TV) as mentioned in
the bug, which is resolved by removing the byte order setting.

b/170529092
b/173631180
b/184060478

Change-Id: Ib2356e9c6a8be47f9a344a14b80df939266573e5

Former-commit-id: 117296d64c4e21c4610c6ee53c2aa07c13a942c5