Disable single test on Android
SbMediaCanPlayMimeAndKeySystem.ValidatePerformance is taking ~1.3ms
on Android. This is long enough that it warrants investigating further.
The test will be enhanced to test all valid A/V formats as well as a
few invalid formats, then the total time of all the formats' tests will
be measured to consider whether we need to optimize this platform,
or adjust the allowed time.

b/184766858

Change-Id: I62111d9de84e789760e7150490ef918a969b6bbd

Former-commit-id: 4114037101aa44d65be3d62f97dc608f0aaf48ad