[ICU] typedef UChar based on wchar_t size
Instead of defining UCHAR_TYPE in the gyp file based on the
platform name, we should typedef UChar directly in the configuration
header file based on whether the platform has a 2 or 4 byte
wchar_t implementation. Keeping this logic in line with one another
will ensure that all platforms get the correct UChar usage.

b/120867447

Change-Id: I359179eeb8e3ef118c096c9ea11443238508b21e

Former-commit-id: 4da910550206991bb0e520dd7cc5e6408d753b67