Starboardize third_party/lz4_lib.
This change Starboardizes the code in third_party/lz4_lib with minimal
changes. The changes to xxhash.c are limited to the beginning of the
file. To starboardize the rest of the code:

1) Following the README.md, amalgamate lz4 code into a single file.
2) Resolve LZ4 memory management leaks with external functions.
3) Resolve xxhash.c memory management leaks inline.

The changes are trivial, and amalgamating the code into a single file
allows us to only modify xxhash.c and leave the rest of the code
untouched. The only change made to lz4_all.c is to prevent the
redefinition of some types.

b/142134121

Change-Id: I067446632b662c19e99cfc9029dfbb17ca9b8d8a

Former-commit-id: 8495951bb5b253e0ef830842b6ffcb9c8130fb5d