[android] Allow to force tunnel mode via constant
Add `kForceTunnelMode` to force tunnel mode when it is set to true.  It
is set to false by default.

Tunnel mode has to be explicitly enabled via mime attribute
"tunnelmode" by the web app.  Set `kForceTunnelMode` to true forces
tunnel mode on all playbacks even if it is not enabled by the web app,
which can be used to:
1. Always enable tunnel mode on certain platforms.
2. Always enable tunnel mode for debugging.

b/175883701

Change-Id: I90baf37bca22935269a3bde106575f18f67891bf

Former-commit-id: 882701b8ea431a0520c389be37ff0b2a7881f4f7