[android] Disable SbMediaConfigurationTest.ValidatePerformance
SbMediaConfigurationTest.ValidatePerformance is taking ~3.9ms
on Android. This is long enough that it warrants investigating further.
The test will be enhanced to test all valid A/V formats as well as a
few invalid formats, then the total time of all the formats' tests will
be measured to consider whether we need to optimize this platform,
or adjust the allowed time.

b/186023317

Change-Id: I1ea6774a04fed70ec0cdc6e368dfdb848fd3c086

Former-commit-id: 6c328a5811529a0fac1afef860d60f4c7cd64e54