[android] Disable SbMediaConfigurationTest.ValidatePerformance
SbMediaConfigurationTest.ValidatePerformance is taking ~3.9ms
on Android. This is long enough that it warrants investigating further.
The test will be enhanced to test all valid A/V formats as well as a
few invalid formats, then the total time of all the formats' tests will
be measured to consider whether we need to optimize this platform,
or adjust the allowed time.

b/186023317

Change-Id: I1ea6774a04fed70ec0cdc6e368dfdb848fd3c086

Former-commit-id: c423816c6ad62f6f42352ee7b089bfb5a5784e1e