Report error on invalid audio configuration
Previously, a DCHECK in starboard_player.cc forced a crash when the
created SbPlayer was invalid. This was not the desired behavior in a
case where we try to create a player with more audio channels than the
platform supports. We now report an error when the player is invalid,
preventing a crash.

b/183965492

Change-Id: I37a085dbcaa76411408b77d9cc5924297569dfa9

Former-commit-id: e73633abb7f827beeab6706101208d70381efb3b