Refactor Windows Docker & Add GCS download steps
This refactor is cosmetic and makes it easier to understand how the
different Dockerfiles relate.

The hierarchy of images and dockerfiles are:

* windows/base/visualstudio2017 : visual-studio-base
  * windows/base/build : build-base-windows
    * windows/win32 : cobalt-build-win32
    * windows/ps4 : cobalt-build-ps4
    * windows/xb1 : cobalt-build-xb1

Additionally, in this CL the PS4 docker build steps are fixed to use GCS
to download the installers and SDK with the service keyfile as an
argument to the build (similar to how XB1) is done.

b/187530874
b/186676115

Change-Id: Id2a24551978374ffcde1f534a92af8b75d2dda8c

Former-commit-id: 0637721c5cfe15524175fcb71e4f400aff216118