Deprecate SbStringCopy, StringCopyWide and StringCopyUnsafe
Note, SbStringCopy is defined and implemented as an equivalent of
BSD non-standard strclpy, which behaves differently than standard
strncpy, in that it always null-terminates.
Most of the reverted code in Chrome upstreams and third_party code
however uses strncpy in original, and takes care of zero-termination,
so the reverts use strncpy too.
In Starboard code, strlcpy implementation is moved to starboard/common
string.h header.

b/149243810

Change-Id: If95b628a7861f50134a80cd7bbb6312d2eacb8ad

Former-commit-id: de6c929cc1cb88f0adc3b53693f7c466260cdc3a