Improve UA conformance for Linux platform.
This adds default values for all fields on the Linux platform, using
the name of the field as the default value.  Use the Starboard
version to derive the Model Year field.

Implement SbSystemGetConnectionType() on Linux by using the netlink
socket API, and searching for the top priority route, and reporting
whether that route uses a wireless network interface.

Adds an nplb unit test to verify that either a wireless or a wired
connection is returned.

b/159647511

Change-Id: Ie9b595abb95ddeb61c557cd11f92856c76ab7d4d

Former-commit-id: 819f762f10f896ad594628159ce5b745402c1daf