Cancel tasks run by update engine on shutdown.
This CL basically attemps to "cancel the running tasks by canceling the
current action update" as mentioned at
https://buganizer.corp.google.com/issues/162114992#comment18. When the
component is in state kDownloading(Diff), the URL request for the
download is canceled.

Existing unit tests for this code are currently disabled and will be
updated / added to in a separate change.

b/162114992

Change-Id: I33684f2c86aa3af47e229d6292197fbab01998ab

Former-commit-id: 1a4b1a2218cc35752effbb2e4be48842e0a464fa