Refine error reporting on invalid audio config
This changes a call to CallSeekCB() from within the scope of the
SbPlayerPipeline |lock_| to outside of it. This resolves the deadlock
issue introduced by go/lbreview/191702 in 21 LTS only; the deadlock
doesn't occur in other branches.

This change replaces a call to CallErrorCB() with CallSeekCB() in
trunk and 22 LTS.

b/193071656
b/183965492

Change-Id: I3f94b66fcf4af33d993f3f97f44f00c2237eb444

Former-commit-id: a5a9ec6a3022db993ec4edbad8b8e009de7b69fc