[android] Refactor AudioSink max frames per request
AudioTrack requires the audio frames to be copied into a Java array
before writing.  The AudioTrackBridge reuses a Java array to avoid
allocating a new Java array for every write.  The size of the Java array
dictates the maximum number of frames that can be written into the
AudioTrack in one write, and it was set via a ctor parameter of
AudioTrackBridge.

All places call the ctor set it to 65536, which is enough for all known
playbacks.  Now it is set inside AudioTrackBridge and can no longer be
customized by the caller.

b/176390025

Change-Id: I59a43b5255e9db6e71bfc44544d1198091c33ed6

Former-commit-id: e3967ca7edafad202bb65a4028f7d2ac51ce1b76