Fix GYP failing to verify VS2017 installation
As part of the GYP build step, we verify the installation of
Visual Studio 2017. However, in the case of docker, the installation is
not a one-to-one match with normal workstation installations of VS2017.

As a result, the GYP step was erroneously saying VS2017 was not
installed. This CL fixes this by using a different indicator (i.e.
checking for a particular installed file, either devenv.exe or
devenv.isolation.ini) to verify the installation for
both docker and dev-workstation contexts.

b/194090898

Change-Id: If6df54405aea69d717ebc055b42daf0c7f6929e7

Former-commit-id: f785bc9e688aa6dce41e1106010d1b291173954f