Prevent deploy builds making symlinks in Docker
Docker is unable to create symlinks for windows containers. As a result,
during the deploy build steps, we need an alternative to copy the files
instead of creating symlinks.

This CL address this for most windows builds, by ensuring we check for
an environment variable |IS_DOCKER| during GYP and Ninja compilation
steps. If it is a windows-docker environment, then symlink steps are
replaced with copies.

b/194208915

Change-Id: I9d608b5507d06b7a8cf3b360f1d29602e03d0672

Former-commit-id: bd3e0303b237fe25cc311f2ca13b91b5866ba1e1