[android] Implement audio passthrough MIME toggle
The MIME parameter "audiopassthrough" is added to allow toggling
passthrough audio. Passthrough audio is enabled by default. If the
player attempts to create a passthrough audio renderer when
'audiopassthrough="false"' and creation_parameters.audio_codec() ==
kSbMediaAudioCodecAc3 or kSbMediaAudioCodecEac3, the player
creation will fail and an error message will be displayed to the
user.

Filters for most e/ac3 tests on Android are removed as tests now specify
if passthrough should be disabled.

b/189781157

Change-Id: Ie65ee9b714f706f40b949807a83fd32e01cda633

Former-commit-id: ec099e752e2050696c47e480232ca8e2d901573e