Adds an example build configuration for Stadia.
This build configuration has been added to
starboard/contrib/stadia/linux/stadia, but it needs to be moved to
starboard/linux/stadia in order to compile.

This configuration is a fork of the starboard/linux/x64x11
configuration.

Also included in this change is a fork of the starboard/shared/x11
application class, which needs to be modified to initialize Stadia
platform services.

b/194504983

Change-Id: Id673ac8d77ed8edad7cf93ca2433039eddc5044e
(cherry picked from commit 923c5793ce70098373ee5f4c0153dc1fe04e0a82)

Former-commit-id: 55b6ee3a9db4f26664b1e5297993c7ab4285b448