Adds an example build configuration for Stadia.
This build configuration has been added to
starboard/contrib/stadia/linux/stadia, but it needs to be moved to
starboard/linux/stadia in order to compile.

This configuration is a fork of the starboard/linux/x64x11
configuration.

Also included in this change is a fork of the starboard/shared/x11
application class, which needs to be modified to initialize Stadia
platform services.

b/194504983

Change-Id: Id673ac8d77ed8edad7cf93ca2433039eddc5044e

Former-commit-id: 39e5ef10b0bc0320505ea33c2a0c1fc78198c355