Fix Camera3D crash with preload
When preloading, the initial Camera3D that was created does
not have an associated impl() (i.e. an input::Camera3D)
because an input manager is not instantiated in preload.

This change allows graceful transition from the placeholder
dom::Camera3D without input to a dom::Camera3D with input.

b/198526160

Change-Id: Ib65355a4205c971f37c5499b599dc877975acadc

Former-commit-id: 825fd5623a5b4010ea4e7f1ed06022f56cbc4d69