Refactor run_evergreen_tests to support MH.
These are the Cobalt-side changes for http://go/coeg-on-mh-impl.

run_evergreen_tests.py should now only be called by build workers and
wraps evergreen_tests.py. evergreen_tests.py is the base script that now
contains most of the logic and is expected to be called directly by
Mobile Harness and individuals, and indirectly by build workers.

Separating the scripts in this way seems more clear and also follows the
pattern used for black box tests. And the open-sourced
starboard/evergreen/testing/README.md now points to a helper script that
is also open-sourced; this was not the case previously because the buildbot
worker scripts aren't released.

b/171314488

Change-Id: Ia8b0198d23368c44f98af097cafe857840562ec2

Former-commit-id: 2385b823b3a5f0e06ae2c282b42eebd3a0af4d2c