Cache fonts within one font family separately.
When each font has it's own mapping, the caches don't get confused if
the character mappings within a family are mismatched. The cache is
populated when the font is scanned, with the drawback being that it uses
approximately 2x the space if all the fonts in the family are being used.

b/211528579

Change-Id: I7bb6f25c6c710c95224dca2070db1f78e2c1c651
(cherry picked from commit b5def92b1a27de3f1913799a45fdd9e2364f2d02)

Former-commit-id: d0f80b847db68a8ee4b67275e7c3fb0a420b9ecd