Cache fonts within one font family separately.
When each font has it's own mapping, the caches don't get confused if
the character mappings within a family are mismatched. The cache is
populated when the font is scanned, with the drawback being that it uses
approximately 2x the space if all the fonts in the family are being used.

b/211528579

Change-Id: I7bb6f25c6c710c95224dca2070db1f78e2c1c651

Former-commit-id: 6ab661230e17fd77287018436735aff42c1c150f