Pass device_id param as dev_servers_listen_ip in black_box_tests.py
When running on MH we are specifying device to use in test by device_id.
As a workaround for b/206119532 and b/206990100 we will use value of
device_id if it is valid IPv4 orIPv6 as --dev_servers_listen_ip param for
cobalt start.

This is workaround for issue when platform is declaring support of IPv6
but failing to bind IPv6 localhost.

Bug: 206119532
Bug: 206990100
Test: Pre-submit run

Change-Id: I80cecccf2574b70882f1151a9706c6312b96e690
(cherry picked from commit 9c90bc184926d655f5eb5cc07b8934d6c7a3c9de)

Former-commit-id: 7cd87384d190c64f7356db1008bfc84247eaebcc