[GN] Migrate evergreen-x64 platform.
This is a first attempt at migrating the platform and additional work
(e.g., toolchain configuration) and testing will be needed.

The goal for now is to be able to use the platform to build some
evergreen-specific dependencies that require the evergreen platform's
configuration. The only existing target that can definitely be built now
for this platform is third_party/musl:c, but this change should allow
the targets under third_party/llvm-project to be built as well.

b/206642994

Change-Id: I9f79f7820147e878d25765a3025e105208fa5e03

Former-commit-id: 105a1dd46a2fe240181f0fdd5e4100248a453d48