[Android]: Fix background memory stay aggressively.
We need to stop MediaPlaybackService before suspending so that this
foreground service would not prevent the activity's memory consumption
getting freed.

There are two ways to stop MediaPlaybackService. One is to stop it
before activity goes to foreground. Second is to stop it before
lifecycle state transits from Concealed to Frozen.

b/207676928

Change-Id: I13aae6f41b3a7303de40a671cb07f633851b7842

Former-commit-id: 087247accf7ed34adfe00956dbcd2c9d353e7eed