[Android]: Fix mediasession is still alive after suspending.
Before this CL, when the mediaSession is released, the activating is
always true which would create a new mediaSession, thus we still have
a alive mediaSession after suspending.

With this CL, we detect the mediaSession is null or not, if it is null
and playbackState == NONE we set the activating to false, which would
not create a new mediaSession after suspending.

b/203731459, b/204597957

Change-Id: Iaf4a2c03837088d0fafbfcf9ce4bb57681b841a8
(cherry picked from commit 45b636b977bf8458bc6816548025ef92e7e45f58)

Former-commit-id: c03a68d377362be9530b934c1037dcf9953cc825