Dedicated Worker stubs.
This adds initial IDL files and stubs for Dedicated Worker support.

There is no actual functionality in this yet, but some of the code
structure is laid out.

This also includes the first steps to connect the MessagePort on the
client side, but some more changes are needed to allow code re-use
for the client and server side interfaces.

b/218791814

Change-Id: Ibb4495afaf38eb4d6386316dff0fdabcce8997a9

Former-commit-id: 552c072f8a8e813edd285adbbc58da7c552cd072