Clarify log about bytes written during compression
LZ4F_compressUpdate() is implemented so that uncompressed input data is
just added to an internal buffer until that internal buffer grows big
enough that it's time to actually compress the data and generate a
block. Because of this behavior and our currenet choices of num_bytes
(input size) and max block size, only ~1 in 30 calls to this function
actually results in bytes getting written to the out buffer maintained
by the LZ4WriterDelegate. This change just updates the log statement to
clarify why it usually says zero bytes will be written.

b/198787888

Change-Id: Ia64c2f46bb5e0a72e1330de7bcd5e2c318f179b8

Former-commit-id: e5b0f6ad2f452115a9cef55ee69e933776a47325