Add ServiceWorkerJobs
This adds a worker::ServiceWorkerJobs that can contain the service
worker registry related algorithms, and the related queueing, tasks,
promises, etc.

The difference between worker::ServiceWorkerJobs and
browser::ServiceWorkerRegistry is that the former can contain the
algorithms following the worker specs, while the latter can handle
the logic that is browser dependent, such as the writing and reading
to/from disk.

Add stubs for the "Start Register" and "Match Service Worker
Registration" algorithms.

Fill in more of ServiceWorkerContainer methods Register() and
GetRegistration() to collect the parameters, and hand off to the
ServiceWorkerJobs.

b/228901874

Change-Id: I6d284a8f5c91d4ff30e22ff3fd3c52dc804a33cc

Former-commit-id: c67499131f7b5d3c616dbfa49341e5107859d00a