Have evergreen tests use a finite expect timeout
This may give use more information about whether the expect dialog and
underlyiing ssh connections are failing when the evergreen tests time
out. 360 seconds is used for now because the longest tests currently
require cobalt to be run for up to 300 seconds.

Also, stop discarding the output of commands in raspi/deploy_cobalt.sh
in case it contains useful information.

b/228980367

Change-Id: I26a7927a45b8b001d43198e7e310fdb1d143940d

Former-commit-id: 1ea87615b38423ea4cab144001cd892c57a52f31