Implement service worker registration getter and setter algorithms.
This implements:
* Algorithm for 'Get Registration'.
* Algorithm for 'Set Registration'.
* Algorithm for 'get the service worker registration object'.
* Algorithm for 'get the service worker object'.
* First steps of algorithm for "Update".

This adds the 'service worker object', the 'service worker
registration object map', the 'service worker object', and 'scope to
registration map'.

This extends the service worker jobs promise rejection to be able to
reject with simple exceptions.

This updates the service worker black box test which now returns a
registration when there is no error.

b/228909717

Change-Id: Ib9539abfc38dcccecdf13933368ff7061f71fadd

Former-commit-id: 9cea49f83562a5c822ceba6a3fd742974a1b386d