Increase sleep time in android launcher from 2 secs to 5.
While troubleshooting why Cobalt Media Integration Tests have been flakey, myself and @***REMOVED*** discovered that the cases where it was failing were when YouTube app was already running on a device, which happens due to Vega not closing YouTube app when test has finished.

There is a CL in flight - cl/447018617 - which should fix this, but
until that gets merged, increase sleep time between killing previous app
and starting new one to see if this stabilizes.

b/232446163

Change-Id: I7bfe2b1698bf5643b8862b1d24e2fb70beb835fb

Former-commit-id: 109f7157338778f706323e51c3df8eeae1ec36a9