Implement Activate and Try Activate algorithms.
This implements the algorithms for Activate, Try Activate, and the
interface WorkerNavigator (with a helper NavigatorBase class to
reduce code duplication).

This implements some smaller algorithms related to service worker
activation, including the algorithm for the ready promise attribute,
the algorithm for controller, and the algorithm for Notify Controller
Change.

This adds some missing event handlers used in window and worker
global scope objects.

Changed the ServiceWorkerObject references in
ServiceWorkerRegistrationObject to scoped_refptr<> so that they can
own the objects, as well as purely follow the algorithms in the spec
for updating the installing, waiting, and active workers that shift
ownership around.

b/228904229

Change-Id: I7297fcfb8dd8d50e9c2f689a1f6a00d4e130e562

Former-commit-id: 6ffb4a42ae78da2fc10499bb19310aa92966273d