Move shared classes from dom to web.
This moves classes shared by Window and Workers from DOM to Web.

This will make it easier to reduce depenency cycles between worker
and window contexts.

To help with that, the web and xhr parts are split out of
dom::GlobalStats into their own classes.

Note: DOMException is also moved to web, even though it has the name
'dom' in it.  That class actually is not part of DOM, since it does
not depend on anything from dom and can be used to reject promises in
worker contexts.

b/229754144

Change-Id: Id782e5a6a4bd7b845c11dabafa61c5d03fc41bc8

Former-commit-id: fa8b90a37ee3a026aeb50b865aa20250ddaf78c7