Window or Worker Global scope.
This adds a WindowOrWorkerGlobalScope.  It doesn't make use of it
yet, but it's there for exposing properties shared beween Window and
Worker global scopes.

A method GetWindowOrWorkerGlobalScope() is added to the web::Context
to get a pointer to the object.

Three added helper methods IsWindow(), IsDedicatedWorker(), and
IsServiceWorker() are available to easily identify the gloval scope.

b/232960199

Change-Id: Ieb918e2b036f28d4dc1642d36c408ade1c86424c

Former-commit-id: c51e1aa593e411f253cc2daa9169be6615069492