Attempt to use extradata to parse codec config for h.264 and h.265.
This code is based on AVStreamToVideoDecoderConfig in chromium's
media/ffmpeg/ffmpeg_common.cc. Since this code cannot depend on chromium
code, we don't parse the full config from the extradata; instead we only
read the first two bytes (the version and an int used to derive the
codec profile).

Bug: b/231631898
Change-Id: I80eefdeb34910b0ba2069c4f5b15d15d1d7c58e3
NOKEYCHECK=True
GitOrigin-RevId: f6e6e50a5597ec53009759d1cb6e0dee9004ab78

Former-commit-id: ebae8629026a8e51fe65ae56f170a26d50ef6431