Allow global constructors to be exposed to workers.
This adds generation of global constructor definitions for interfaces
exposed to worker gloal scope web contexts.

This uses that for Event and ErrorEvent, and a unit test is added to
verify that the constructor is available in WorkerGlobalScope as part
of a unit test for the onerror global event handler.

b/234186834

Change-Id: Ia644f387b26e197d6c39ffdedc85e3c70c954a8e

Former-commit-id: cd711f7dde8a1328f66b2f14ac49032beeb0c9fe