Move UserAgentPlatformInfo from WebModule to BrowserModule.
The retrieval of the UserAgentPlatformInfo is moved from WebModule to
BrowserModule, so that it can be re-used between WebModule and Worker
instances, and so that it can be accessed from the web Context in the
NavigatorBase constructor, avoiding the creating of separate object
instances.

Removed counting of the onchange event in the service worker black
box text because it's not deterministic and causes flakyness.

b/233951330

Change-Id: Idee24fb405b556e72a7935f932a9bc302cdc382d

Former-commit-id: a712d069fee06f6523d8cb05e42f9f6c34f4d9fa