Optimize and refactor LZ4 file decompression
The first optimization is the one described in b/233936269. This reduced
average libcobalt load time on a raspi-2 from 544ms to 527ms. The number
of read-decompress iterations was reduced about 10x, from just over 1000
to just over 100, but the impact on decompression time is much smaller.

The second optimization is to use the content size information in the
LZ4 frame header to allocate exacatly as much memory as needed to store
the decompressed data. This is instead of estimating and resizing. On a
raspi-2 this further reduced average libcobalt load time to 511 ms, and
it also reduced memory use during decompression from 78MB to 72.

b/233936269

Change-Id: I4a260ce68cea96aa8b539a3df1bd23ee4b4966c0

Former-commit-id: 7a35c019563a8f6017831a15923b0fba934309f0