Implement terminate algorithms.
This implements the algorithms for terminating web workers and service workers, including the algorithm for Terminate Service Worker.
The ServiceWorkerRegistrationObject is made to be refcounted with
thread safety, so that it's lifetime can be extended until no more
tasks are referring to it.

Some other cleanups are included, such as some more thread checks and
tracing, logging cleanups and adding bug numbers to TODO's.

b/228905127
b/229767497
b/229128880

Change-Id: Iab5026933ffb5a025adb8f62294839ef2c016055

Former-commit-id: 19d9cdc094ed5e89eab01d4ae07459dd40c61181