Create default test sharding configuration
This change will ensure that any platform that is invoked with sharding
will use a default configuration to allocate their tests into multiple
shards (if the corresponding sharding flag is passed).

Note: for platforms that don't fully support sharding, it is necessary
to ensure they are not passed the `--shard_index` flag by the
|test_runner| and |run_unit_tests| python scripts that kick of the
tests.

b/233108244

Change-Id: I1b9659583ffd40bed50328a43f2974246cfe66ec

Former-commit-id: bc4ac4167be4701c6483cc1dbfc9a00222818df7