Implement stubs for interfaces Client and Clients.
This implements stubs for the Client and Clients interfaces, as well
as the getter for Clients in the ServiceWorkerGlobalScope.  All
methods of Clients return a promise that gets rejected with
'NotSupportedError'.

Basic functionality is tested in the service worker test black box
test, although only runtime errors will be tested, because there
isn't a functional MessagePort yet between service workers and their
clients. This change is the first step to making those possible.

b/234731396

Change-Id: Ie7929bc4c90b71d572794fa87ae06bc886e2d17a

Former-commit-id: 4b5a0aa855caf8968f54b1f67ce9494052b77717