Web API implementations for Client and Clients.
This implements methods for the Client and Clients Web APIs including
Clients.matchAll(options) and Clients.get(id).  This also adds some
related functionality such as the environment id value.

Tracking of the active service worker is moved from the
EnvironmentSettings to the web::Context, so that it can be held with
reference counting, to ensure the ServiceWorkerObject survives when
unregistered but still active for a service worker client.  For
similar reasons, a reference to the containing service worker
registration of the active service worker is held in the web::Context
as long as it holds the active service worker.

Registering and Unregistering of web::Contexts is moved from the
web::Context initialization and destructor to the web::Agent to avoid
a reference counting loop prevening shutdown of web::Contexts.

b/234731396

Change-Id: I8eaa40294cbe4a9f527249f3548f446f91924aae

Former-commit-id: 40f8fb786bdf4502cd2f5ce365454365c2b89ccf