Stubs for ExtendableEvent and ExtendableMessageEvent.
This adds stubs for ExtendableEvent and ExtendableMessageEvent, as
well as stubs for FromJSValue for Promise<T>, needed to use a promise
as a parameter.

This also adds the relevant binding generator plumbing, and changes
the script::Promise<T> class to not be pure virtual, because that is
needed for the constructing the base class as an element to pass the
address of to the FromJSValue interface.

This also moved MessagePort from worker to web, even though that
wasn't needed for this CL (yet): MessagePort objects are used in APIs
for both workers and window contexts).

b/228976500

Change-Id: Id6d86ac4a11e0574ab8d77670616c95ec2888c7a

Former-commit-id: 7611bf835b18ec16e1c6831f9b4e2e6d67b49407