Fix threading for a promise and service worker object map maintenance.
This fixes the thread in which the promise is resolved in Clients.get()
when there are no clients.

This fixes the maintenance of the service worker object map that
would cause a ServiceWorker created for a ServiceWorkerObject that
has already been destroyed to still be used if a new
ServiceWorkerObject had later been constructed with the same address.

This adds a virtual destructor to EventTarget.

This adds thread checkers to event targets and promises to allow
catching future threading issues sooner.

b/219986442

Change-Id: Ia0f6c4ac12fca44c5d0a123f84aab9d1f3d0de64

Former-commit-id: 026f7e558e40b9424cb48ff6c88cb71b322d39a9