Optimize LZ4 prefs for size & decompression speed
I tried different combinations of the LZ4 preferences on 1) a Raspberry
Pi 2 and 2) a Chromecast with Google TV (using go/lbreview/225565 to run
Evergreen), and the results are in go/lz4-tuning. The combination in
this CL is the only one that was within 3% of the lowest found a)
libcobalt.lz4 size and b) decompression speed for both devices.

The preferences don't significantly impact memory used during
decompression except for block size when it's increased to 4MB.

We should be able to update most of these preferences in the future
without breaking client-side decompression, so this choice isn't
binding.

b/233939346

Change-Id: If35a3ec05c611cac284f829d989dee716a701a5e
(cherry picked from commit 1b709c1631e3c8543613a1bfe7c36b2f33cd6e2c)

Former-commit-id: b7c47d410b10384df7ffa310eee83f67ffa79123