Optimize LZ4 prefs for size & decompression speed
I tried different combinations of the LZ4 preferences on 1) a Raspberry
Pi 2 and 2) a Chromecast with Google TV (using go/lbreview/225565 to run
Evergreen), and the results are in go/lz4-tuning. The combination in
this CL is the only one that was within 3% of the lowest found a)
libcobalt.lz4 size and b) decompression speed for both devices.

The preferences don't significantly impact memory used during
decompression except for block size when it's increased to 4MB.

We should be able to update most of these preferences in the future
without breaking client-side decompression, so this choice isn't
binding.

b/233939346

Change-Id: If35a3ec05c611cac284f829d989dee716a701a5e

Former-commit-id: 4c02987374b5ed7715744fc1761616a219f62153