Revert "Define a demuxer cobalt extension."
This reverts commit e9d628986f6dc7ef8ec3a4a909d1373d6fa3d0d6.

Reason for revert: This CL breaks the build by relying on a system-provided avformat.h. Builds should not rely on the system having libavformat-dev providing a avformat.h, but should use the matching version from third_party/ffmpeg_includes, as done in the audio and video decoders (which was compiled multiple times, once for each supported ffmpeg version).

Change-Id: Ie0a00e7838a08c0ee8a67f227eba5219ac81a741

Former-commit-id: 31fd001d7d2c13e8a065d391ed87622448e04cfa