Implement WaitUntil and send ExtendableEvent for install and activate.
This implements ExtendableEvent.waitUntil(), and sends that type of
event for the install and activate events.

This extends the promise abstraction to allow JavaScript promises to
be passed to native code, as well as the registration of a native
callback that is called when the state of a promise changes.

Note: This does not complete the logic for ExtendableEvent, there are
additional related TODOs.

b/228976500

Change-Id: I190ecb154796ebea5e0ae0a9812bd3c334ef7ab6
(cherry picked from commit 12d061c27861b8a648a1334970ed913d295553f6)
NOKEYCHECK=True
GitOrigin-RevId: 3a63e0f5534414613ca802c6d57d06bd2c6053c4

Former-commit-id: ad26d02b25e461616f22620f482b205ce987425f