Add an include path to ffmpeg_demuxer_test for ffmpeg includes.
This adds an include path for ffmpeg includes so that it doesn't rely
on the system providing them.

Before and after this CL, ffmpeg_demuxer_test will build only against
one particular version of the ffmpeg includes.  Before this CL that
was the version correponding to the system installed include files.
After this CL, that is the version explicitly added in the build
config.

That means that the ffmpeg_demuxer_test binary will only work well
when run on a machine with the given version of ffmpeg installed.

This test binary actually should either have specializations for all
supported ffmpeg versions, or (preferred) it should only test the
behavior of the abstraction layer without testing implementation
details.

b/219074274

Change-Id: I10c89ab9beef787a05370d3294889167a717fdac

Former-commit-id: 317221ab0801df0bf6938ec024aba4ce8297ca5d