Run Evergreen Lite test w/ compressed system image
Having the automated tests run this test case in this mode should give
us confidence that an LZ4-compressed system image can be successfully
loaded.

A more lightweight implementation, without changes in evergreen_tests.py
or evergreen/shared/launcher.py, would be to add a separate test case
that swaps in libcobalt.lz4 for libcobalt.so on the device and then runs
the Evergreen Lite test. But this would be kind of hacky since the test
would need to somehow bypass the launcher framework and independently
get libcobalt.lz4 from the out/install dir to the device. I think the
implementation in this CL is cleaner and also more extensible: once we
have compressed packages on Omaha/Lorry we also want to test that they
can be installed and upgraded to, and this can simply be done by
enabling another test case when USE_COMPRESSED_SYSTEM_IMAGE=true.

b/233934796

Change-Id: I2f6e74723485a655e0528b63256efebfa58d81cf
NOKEYCHECK=True
GitOrigin-RevId: 96c13cd1127d72620cbb41490f6eaf5c648c3c87

Former-commit-id: 7709907824c401a39ee2afc3990e88a5149db03b