Adding more ExtendableEvent logic.
This adds waiting for the ExtendableEvent to become inactive during
the Install and Activate algorithms.

This also adds some checks that prevent crashes if event handler code
of an ExtendableEvent uses promises from Clients.claim(),
Clients.get(), Clients.matchAll(), or self.skipWaiting(), and they
race termination of the service worker (for example when they are
used incorrectly without a waitUntil()).

b/228976500

Change-Id: I400fa2bd99185aba2cc9749ff2d2ba91c6997a2f
NOKEYCHECK=True
GitOrigin-RevId: d5deebbbf0596f083b104eddf34d5d601959080d

Former-commit-id: ea2d15b2ae7e1b594b516e196c6fb1ef461c79e8